requirements.txt
pytest-mock==3.15.1
pytest-asyncio==1.3.0
pytest==9.0.2
pytest-cov==7.0.0
mcp==1.24.0
pyodbc==5.3.0
python-dotenv==1.2.1